requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
argon2-cffi~=25.1.0
pymongo~=4.15.3
ecdsa[gmpy2]
